// src/interfaces/HomePageInterface.ts
export interface HomepageState {
    imageNamePopularMovies: string;
    isFlippedPopularMovies: boolean;
    isFlippedTopRatedMovies: boolean;
    isFlippedUpcomingMovies: boolean;
    isFlippedNowPlayingMovies: boolean;
  
    imageNameUpcomingMovies: string;
    CardBackTitlePopularMovies: string;
    CardBackTitleTopRatedMovies: string;
    imageNameTopRatedMovies: string;
    CardBackTitleUpcomingMovies: string;
    imageNameNowPlayingMovies: string;
    CardBackNowPlayingMovies: string;
  }